refactor: ProductController 패키지 이동
